+++ index.js
@@ -35,8 +35,8 @@ var spaces = {
 	lchab: require('./lchab'),
 	luv: require('./luv'),
 	lchuv: require('./lchuv'),
-	husl: require('./husl'),
-	huslp: require('./huslp'),
+	hsluv: require('./hsluv'),
+	hsluvp: require('./hsluvp'),
 	cubehelix: require('./cubehelix'),
 	coloroid: require('./coloroid'),
 	hcg: require('./hcg'),
